src/client/requirements.txt
--index-url https://pypi.org/simple/

pyyaml==6.0.1
argparse==1.4.0
requests==2.31.0
